git-svn-id: http://my.redmine.jp/svn/iric/v3.X/gui/trunk@5800 2cc7cdd0-1db9-4218-aa4a-1d8dad43e0f0

+++ libs/misc/xmlsupport.cpp
@@ -83,7 +83,7 @@ int iRIC::getIntAttribute(const QDomNode& node, const QString& name, int default
 	bool ok;
 	double intVal = val.toInt(&ok);
 	if (! ok || std::isnan(intVal)) {intVal = defaultVal;}
-	return intVal;
+	return static_cast<int> (intVal);
 }
 
 void iRIC::setIntAttribute(QXmlStreamWriter& writer, const QString& name, int value)
